move common data classes to base
